bids_wrapper.py: longitudinal copying fix
when running 'shutil.move', and the folder a src file is being moved to doesn't exist,
this would error out. this fix simply makes the bids categorical directory (e.g. fmap/)
for the file if it doesn't already exist to avoid this from happening.

this can happen if, say, the first session in a longitudinal sequence of 3 only contained
anatomical images, whereas sessions 2 and 3 contain functional runs and fieldmaps. the old
behavior would have been to attempt to move these func/fmap files into a folder that didn't
yet exist, but should no longer happen