FEAT: --allow_uneven_fmap_groups flag added
this flag will not error out if there are more/less AP- than PA-oriented fieldmaps, and will
instead make N pairs of AP/PA fieldmaps (with N being the number of whichever fieldmap type
is fewer) based on the fieldmaps' vicinity in time to one another.

for example, if there are 2 AP fieldmaps followed by 1 PA fieldmap in a localizer block,
the second AP and only PA fieldmap will be mapped to the same BOLD run, if the pair has
the closest acquisition time to the BOLD run.